requirements-minimal.txt
# Ultra-minimal requirements - only pre-compiled wheels
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
python-multipart==0.0.6
